+++ src/components/Groups/index.css
@@ -1,24 +1,1 @@
-.cio-filter-groups-options-list ul {
-  padding-left: 0;
-  margin-top: 0;
-}
-
-.cio-filter-groups-options-list .cio-crumb {
-  font-size: 14px;
-  background: none;
-  color: inherit;
-  border: none;
-  font-family: inherit;
-  cursor: pointer;
-  outline: inherit;
-  padding: 2px 4px;
-  border-radius: 4px;
-
-  &:hover:not(:last-child) {
-    background-color: #f2f2f2;
-  }
-}
-
-.cio-filter-groups-options-list .cio-crumb:last-child {
-  cursor: auto;
-}
+@import './Groups.css';
